llm-cloud-run/requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform
google-cloud-storage
pydantic
python-multipart
google-cloud-bigquery
google-cloud-core
